// superlista-front/src/openapi/model/menuSemanalDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DiaComidaDto } from './diaComidaDto';


export interface MenuSemanalDto { 
    id?: number;
    usuarioId?: number;
    diasComidas?: Array<DiaComidaDto>;
}